feature: Add an into() bevy mesh, example viewer.
Add a cargo feature "into_bevy_mesh" that will provide a `From`
implementation so generators like `Cube` can be turned into a bevy `Mesh`.
This doesn't add any dependencies on the default features. It does add
dependencies on `bevy_render` and `glam` when the feature is enabled.

Add an example called "viewer" that uses this feature to view the
generators. It's a small bevy app with some camera control. I did
discover a bug in a cone normal so plus one for visualization.